requirements.txt
asgiref==3.4.1
click==8.0.1
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.0.2
psycopg2-binary==2.9.1
django-environ==0.7.0